+++ TablePro/Models/Query/QueryTabState.swift
@@ -12,6 +12,7 @@ enum TabType: Equatable, Codable, Hashable {
     case createTable      // Create new table tab
     case erDiagram        // ER diagram tab
     case serverDashboard  // Server dashboard tab
+    case terminal         // Embedded database CLI terminal tab
 }
 
 /// Minimal representation of a tab for persistence